fix: change order for withdrawal
