+++ internal/app/handler/order.go
@@ -63,34 +63,6 @@ func (h *OrderHandler) UploadOrder(w http.ResponseWriter, r *http.Request) {
 	w.WriteHeader(http.StatusAccepted)
 }
 
-const lastDigit = 9
-
-func ValidateOrderNumber(number string) bool {
-	var sum int
-	digits := make([]int, len(number))
-	for i, char := range number {
-		if char < '0' || char > '9' {
-			return false
-		}
-		digits[i] = int(char - '0')
-	}
-
-	double := false
-	for i := len(digits) - 1; i >= 0; i-- {
-		digit := digits[i]
-		if double {
-			digit *= 2
-			if digit > lastDigit {
-				digit -= lastDigit
-			}
-		}
-		sum += digit
-		double = !double
-	}
-
-	return sum%10 == 0
-}
-
 func (h *OrderHandler) GetOrders(w http.ResponseWriter, r *http.Request) {
 	userID, ok := r.Context().Value(domain.ContextKey).(int)
 	if !ok {
