+++ internal/domain/entity/withdrawals.go
@@ -1,7 +1,7 @@
 package entity
 
 type Withdrawal struct {
-	Order       string  `db:"number" json:"order"`
+	Order       string  `db:"order" json:"order"`
 	ProcessedAt string  `db:"processed_at" json:"processed_at"`
 	Sum         float64 `db:"sum" json:"sum"`
 }
